chore: move SliderEntity to lgl
