+++ src/main/java/com/logandhillon/fptgame/resource/Colors.java
@@ -18,15 +18,12 @@ public final class Colors {
     public static final Color FOREGROUND_TRANS_40 = Color.rgb(0, 0, 0, 0.4);
     public static final Color FOREGROUND_TRANS_50 = Color.rgb(0, 0, 0, 0.5);
 
-    public static final Color GENERIC_BG = Color.web("#D2D1D1");
+    public static final Color BUTTON_NORMAL = Color.rgb(207, 209, 235);
+    public static final Color BUTTON_HOVER  = Color.rgb(75, 150, 249);
 
-    public static final Color BUTTON_NORMAL      = Color.rgb(207, 209, 235);
-    public static final Color BUTTON_HOVER       = Color.rgb(75, 150, 249);
-    public static final Color SLIDER_HEAD        = Color.rgb(217, 217, 217);
-    public static final Color SLIDER_HEAD_ACTIVE = Color.rgb(166, 172, 194);
+    public static final Color PLAYER_RED  = Color.web("#E23B36");
+    public static final Color PLAYER_BLUE = Color.web("#4184E8");
 
-    public static final Color       PLAYER_RED   = Color.web("#E23B36");
-    public static final Color       PLAYER_BLUE  = Color.web("#4184E8");
     /**
      * The color of the player skin, indexed by the order they appear on the main menu.
      */
